% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_numeric.R
\name{match_numeric}
\alias{match_numeric}
\title{Matches groups by using numeric variables}
\usage{
match_numeric(df, n = 10, list = NULL)
}
\arguments{
\item{df}{data frame of numeric inputs. First column has group names, 1 line per group.}

\item{n}{size of the test group, and matching control group.}
}
\value{
Outputs a list of test groups with matching control groups, 1 to 1 match.
}
\description{
Matches groups by using numeric variables
}
