% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_numeric.R
\name{match_numeric}
\alias{match_numeric}
\title{Match_Numeric}
\usage{
match_numeric(input_data, n = 10)
}
\arguments{
\item{n}{size of the test group, and matching control group.}

\item{df}{data frame of numeric inputs. First column has group names, 1 line per group.}
}
\value{
List of test groups with matching control groups, 1 to 1 match.
}
\description{
Matches groups by using numeric variables
}
